Address PR comments for SANS system tests
Split up various system tests into their own files where multiple
instruments previously were
Fix some unrequired imports
Re-org some missed files into the correct dirs
