+++ Testing/SystemTests/tests/framework/ISIS/SANS/SANS2D/SANS2DMultiPeriod.py
@@ -9,7 +9,6 @@
 # test batch mode with sans2d and selecting a period in batch mode
 import systemtesting
 from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
-from mantid.api import AnalysisDataService
 from ISISCommandInterface import *
 from SANSBatchMode import *
 from sans.common.enums import SANSInstrument
@@ -56,55 +55,3 @@ class SANS2DMultiPeriodBatch(SANS2DMultiPeriodSingle):
 
         BatchReduce(csv_file, 'nxs', saveAlgs={})
         self.reduced = '5512_SANS2DBatch'
-
-
-@ISISSansSystemTest(SANSInstrument.LARMOR)
-class LARMORMultiPeriodEventModeLoading(systemtesting.MantidSystemTest):
-    """
-    This test checks if the positioning of all workspaces of a
-    multi-period event-type file are the same.
-    """
-    def __init__(self):
-        super(LARMORMultiPeriodEventModeLoading, self).__init__()
-        self.success = True
-
-    def _get_position_and_rotation(self, workspace):
-        instrument = workspace.getInstrument()
-        component = instrument.getComponentByName("DetectorBench")
-        position = component.getPos()
-        rotation = component.getRotation()
-        return position, rotation
-
-    def _clean_up(self, base_name, number_of_workspaces):
-        for index in range(1, number_of_workspaces + 1):
-            workspace_name = base_name + str(index)
-            monitor_name = workspace_name + "_monitors"
-            AnalysisDataService.remove(workspace_name)
-            AnalysisDataService.remove(monitor_name)
-        AnalysisDataService.remove("80tubeCalibration_18-04-2016_r9330-9335")
-
-    def _check_if_all_multi_period_workspaces_have_the_same_position(self, base_name, number_of_workspaces):
-
-        reference_name = base_name + str(1)
-        reference_workspace = AnalysisDataService.retrieve(reference_name)
-        reference_position, reference_rotation = self._get_position_and_rotation(reference_workspace)
-        for index in range(2, number_of_workspaces + 1):
-            ws_name = base_name + str(index)
-            workspace = AnalysisDataService.retrieve(ws_name)
-            position, rotation = self._get_position_and_rotation(workspace)
-            self.assertEqual(position, reference_position)
-            self.assertEqual(rotation, reference_rotation)
-
-    def runTest(self):
-        LARMOR()
-        Set1D()
-        Detector("DetectorBench")
-        MaskFile('USER_Larmor_163F_HePATest_r13038.txt')
-        AssignSample('13038')
-        base_name = "13038_sans_nxs_"
-        number_of_workspaces = 4
-        self._check_if_all_multi_period_workspaces_have_the_same_position(base_name, number_of_workspaces)
-        self._clean_up(base_name, number_of_workspaces)
-
-    def validate(self):
-        return self.success
