+++ Testing/SystemTests/tests/framework/ISIS/SANS/SANS2D/SANS2DSlicing.py
@@ -8,7 +8,6 @@
 import systemtesting
 from ISIS.SANS.isis_sans_system_test import ISISSansSystemTest
 
-from mantid.kernel import config
 from mantid.simpleapi import mtd, RenameWorkspace, FileFinder
 import ISISCommandInterface as i
 from sans.common.enums import SANSInstrument
@@ -18,7 +17,6 @@ from sans.common.enums import SANSInstrument
 class SANS2DMinimalBatchReductionSliced(systemtesting.MantidSystemTest):
     def __init__(self):
         super(SANS2DMinimalBatchReductionSliced, self).__init__()
-        config['default.instrument']='SANS2D'
 
     def runTest(self):
         import SANSBatchMode as batch
